+++ src/service_identity/__init__.py
@@ -7,6 +7,7 @@ from __future__ import absolute_import, division, print_function
 from . import pyopenssl
 from .exceptions import (
     CertificateError,
+    SubjectAltNameWarning,
     VerificationError,
 )
 
@@ -21,11 +22,12 @@ __author__ = "Hynek Schlawack"
 __email__ = "hs@ox.cx"
 
 __license__ = "MIT"
-__copyright__ = "Copyright (c) 2014-2015 Hynek Schlawack"
+__copyright__ = "Copyright (c) 2014 Hynek Schlawack"
 
 
 __all__ = [
     "CertificateError",
+    "SubjectAltNameWarning",
     "VerificationError",
     "pyopenssl",
 ]