+++ src/service_identity/exceptions.py
@@ -1,5 +1,5 @@
 """
-All exceptions thrown by service_identity.
+All exceptions and warnings thrown by ``service_identity``.
 
 Separated into an own package for nicer tracebacks, you should still import
 them from __init__.py.
@@ -8,6 +8,14 @@ them from __init__.py.
 from __future__ import absolute_import, division, print_function
 
 import attr
+
+
+class SubjectAltNameWarning(Warning):
+    """
+    Server Certificate does not contain a ``SubjectAltName``.
+
+    Hostname matching is performed on the ``CommonName`` which is deprecated.
+    """
 
 
 @attr.s