+++ tests/test_pyopenssl.py
@@ -1,13 +1,15 @@
 from __future__ import absolute_import, division, print_function
 
-from OpenSSL.test.util import TestCase
+import pytest
 
 from service_identity._common import DNSPattern, URIPattern
+from service_identity import SubjectAltNameWarning
 from service_identity.pyopenssl import extract_ids, verify_hostname
+
 from .util import CERT_CN_ONLY, CERT_DNS_ONLY, CERT_OTHER_NAME
 
 
-class VerifyHostnameTestCase(TestCase):
+class TestVerifyHostname(object):
     def test_verify_hostname(self):
         """
         It's just a convenience one-liner.  Let's check it doesn't explode b/c
@@ -20,33 +22,33 @@ class VerifyHostnameTestCase(TestCase):
         verify_hostname(FakeConnection(), u"twistedmatrix.com")
 
 
-class ExtractIDsTestCase(TestCase):
+class TestExtractIDs(object):
     def test_dns(self):
         """
         Returns the correct DNSPattern from a certificate.
         """
         rv = extract_ids(CERT_DNS_ONLY)
-        self.assertEqual(
-            [DNSPattern(b'www.twistedmatrix.com'),
-             DNSPattern(b'twistedmatrix.com')],
-            rv
-        )
+        assert [
+            DNSPattern(b"www.twistedmatrix.com"),
+            DNSPattern(b"twistedmatrix.com")
+        ] == rv
 
     def test_cn_ids_are_used_as_fallback(self):
         """
-        CNs are returned as DNSPattern if no other IDs are present.
+        CNs are returned as DNSPattern if no other IDs are present
+        and a warning is raised.
         """
-        rv = extract_ids(CERT_CN_ONLY)
-        self.assertEqual(
-            [DNSPattern(b'www.microsoft.com')], rv
-        )
+        with pytest.warns(SubjectAltNameWarning):
+            rv = extract_ids(CERT_CN_ONLY)
+        assert [
+            DNSPattern(b"www.microsoft.com")
+        ] == rv
 
     def test_uri(self):
         """
         Returns the correct URIPattern from a certificate.
         """
         rv = extract_ids(CERT_OTHER_NAME)
-        self.assertEqual(
-            [URIPattern(b'http://example.com/')],
-            [id for id in rv if isinstance(id, URIPattern)]
-        )
+        assert [
+            URIPattern(b"http://example.com/")
+        ] == [id for id in rv if isinstance(id, URIPattern)]